Works, it's training

+++ torchgeo/datasets/oscd.py
@@ -409,7 +409,7 @@ class OSCDDataModule(pl.LightningDataModule):
     def preprocess(self, sample: Dict[str, Any]) -> Dict[str, Any]:
         """Transform a single sample from the Dataset."""
         sample["image"] = sample["image"].float()
-        sample["mask"] = sample["mask"].float()
+        sample["mask"] = sample["mask"]
         sample["image"] = self.norm(sample["image"])
         sample["image"] = torch.flatten(  # type: ignore[attr-defined]
             sample["image"], 0, 1
@@ -432,11 +432,11 @@ class OSCDDataModule(pl.LightningDataModule):
         def n_random_crop(sample: Dict[str, Any]) -> Dict[str, Any]:
             images, masks = [], []
             for i in range(self.num_patches_per_tile):
-                mask = repeat(sample["mask"], "h w -> t h w", t=2)
+                mask = repeat(sample["mask"], "h w -> t h w", t=2).float()
                 image, mask = self.rcrop(sample["image"], mask)
                 mask = mask.squeeze()[0]
                 images.append(image.squeeze())
-                masks.append(mask)
+                masks.append(mask.long())
             sample["image"] = torch.stack(images)
             sample["mask"] = torch.stack(masks)
             return sample
